Support `-fno-exceptions`.
The only thing preventing the non-exception portions of GSL from
being used with `-fno-exceptions` is `gsl::narrow`. If we allow the
`-fno-exceptions` case to terminate with `gsl::Ensures` instead, the
library becomes usable in this case.

Note that AFAIK we can only automatically detect this case for Clang.
Non-Clang users will need to define `GSL_NO_EXCEPTIONS` themselves
when building.